валидация

+++ KR/Models/Car.cs
@@ -10,9 +10,9 @@ namespace KR.Models
     {
         public int CarId { get; set; }
         public string Model { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Doors { get; set; }
-        public int Consumption { get; set; }
+        public decimal Consumption { get; set; }
         public string Description { get; set; }
         public string Fuel { get; set; }
